+++ src/util/BookmarkFile.cpp
@@ -46,7 +46,7 @@ bool BookmarkFile::load(const std::string& bookPath, std::vector<BookmarkEntry>&
 bool BookmarkFile::save(const std::string& bookPath, const std::vector<BookmarkEntry>& bookmarks) {
   for (const auto& bookmark : bookmarks) {
     if (bookmark.name.size() > BookmarkEntry::MAX_NAME_LENGTH) {
-      LOG_ERR("BKM", "Bookmark name exceeds %zu bytes", BookmarkEntry::MAX_NAME_LENGTH);
+      LOG_ERR("BKM", "Bookmark name exceeds %lu bytes", static_cast<unsigned long>(BookmarkEntry::MAX_NAME_LENGTH));
       return false;
     }
   }
